+++ Sources/Sentry/SentryEvent.m
@@ -1,5 +1,6 @@
 #import "NSDate+SentryExtras.h"
 #import "NSDictionary+SentrySanitize.h"
+#import "SentryANRTrackingIntegration.h"
 #import "SentryBreadcrumb.h"
 #import "SentryClient.h"
 #import "SentryCurrentDateProvider.h"
@@ -200,6 +201,12 @@ NS_ASSUME_NONNULL_BEGIN
 }
 
 #endif // SENTRY_HAS_METRIC_KIT
+
+- (BOOL)isAppHangEvent
+{
+    return self.exceptions.count == 1 &&
+        [self.exceptions.firstObject.type isEqualToString:SentryANRExceptionType];
+}
 
 @end