+++ Sources/Sentry/SentryFileManager.m
@@ -42,7 +42,7 @@ createDirectoryIfNotExists(NSString *path, NSError **error)
  * from.
  */
 void
-_unsafe_removeFileAtPath(NSString *path)
+_non_thread_safe_removeFileAtPath(NSString *path)
 {
     NSError *error = nil;
     NSFileManager *fileManager = [NSFileManager defaultManager];
@@ -307,7 +307,7 @@ SentryFileManager ()
 - (void)removeFileAtPath:(NSString *)path
 {
     @synchronized(self) {
-        _unsafe_removeFileAtPath(path);
+        _non_thread_safe_removeFileAtPath(path);
     }
 }
 
@@ -794,7 +794,7 @@ writeAppLaunchProfilingConfigFile(NSMutableDictionary<NSString *, NSNumber *> *c
 void
 removeAppLaunchProfilingConfigFile(void)
 {
-    _unsafe_removeFileAtPath(launchProfileConfigFileURL().path);
+    _non_thread_safe_removeFileAtPath(launchProfileConfigFileURL().path);
 }
 #endif // SENTRY_TARGET_PROFILING_SUPPORTED