+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Cache installationID async to avoid file IO on the main thread when starting the SDK (#3601)
 
+### Fixes
+
+- Finish TTID span when transaction finishes (#3610)
+- Dont take screenshot and view hierarchy for app hanging (#3620)
 
 ## 8.20.0